# Further improve the performance of the navigational mesh