We conserved memory by replacing the reference of the Amphibious grid and the Hover grid with the Land grid if we learned they are equal statistically. This happens when the map has no water. This was done at the end of the generation phase. We now do this equality check as soon as possible to allow us to skip various computations on the amphibious and hover grids. This can result in up to 50% faster generation on land only maps. The effect is less significant on larger maps because the dominant factor is not the post processing of the grid but the number of height samples that we need to process to construct the grid.

Old
```
Ian's Cross:
DEBUG: NavGenerator - Flattened 836 sections
DEBUG: NavGenerator - culled 1512 labels
DEBUG: Hover grid equals land grid - ditching hover grid
DEBUG: Amphibious grid equals land grid - ditching amphibious grid
DEBUG: Generated navigational mesh in 3.422852 seconds
DEBUG: Allocated megabytes for navigational mesh: 13.228302
DEBUG: Allocated megabytes for labels: 0.510406

Open Palms: FAF version
DEBUG: NavGenerator - Flattened 908 sections
DEBUG: NavGenerator - culled 102 labels
DEBUG: Hover grid equals land grid - ditching hover grid
DEBUG: Amphibious grid equals land grid - ditching amphibious grid
DEBUG: Generated navigational mesh in 1.071777 seconds
DEBUG: Allocated megabytes for navigational mesh: 4.909599
DEBUG: Allocated megabytes for labels: 0.080338

Syrtis Major - FAF version:
DEBUG: NavGenerator - Flattened 857 sections
DEBUG: NavGenerator - culled 312 labels
DEBUG: Hover grid equals land grid - ditching hover grid
DEBUG: Amphibious grid equals land grid - ditching amphibious grid
DEBUG: Generated navigational mesh in 1.595215 seconds
DEBUG: Allocated megabytes for navigational mesh: 7.185905
DEBUG: Allocated megabytes for labels: 0.131805
```

New

```
Ian's Cross:
DEBUG: NavGenerator - Flattened 836 sections
DEBUG: NavGenerator - replacing amphibious grid with land grid to conserve memory
DEBUG: NavGenerator - replacing hover grid with land grid to conserve memory
DEBUG: NavGenerator - culled 504 labels
DEBUG: Generated navigational mesh in 1.794434 seconds
DEBUG: Allocated megabytes for navigational mesh: 13.228302
DEBUG: Allocated megabytes for labels: 0.197205

Open Palms - FAF version:
DEBUG: NavGenerator - Flattened 908 sections
DEBUG: NavGenerator - replacing amphibious grid with land grid to conserve memory
DEBUG: NavGenerator - replacing hover grid with land grid to conserve memory
DEBUG: NavGenerator - culled 34 labels
DEBUG: Generated navigational mesh in 0.519043 seconds
DEBUG: Allocated megabytes for navigational mesh: 4.909599
DEBUG: Allocated megabytes for labels: 0.055145

Syrtis Major - FAF version:
DEBUG: NavGenerator - Flattened 857 sections
DEBUG: NavGenerator - replacing amphibious grid with land grid to conserve memory
DEBUG: NavGenerator - replacing hover grid with land grid to conserve memory
DEBUG: NavGenerator - culled 104 labels
DEBUG: Generated navigational mesh in 0.686523 seconds
DEBUG: Allocated megabytes for navigational mesh: 7.185905
DEBUG: Allocated megabytes for labels: 0.063507
```